+++ src/main/java/nl/daanh/hiromi/CommandManager.java
@@ -3,8 +3,8 @@ package nl.daanh.hiromi;
 import net.dv8tion.jda.api.events.interaction.ButtonClickEvent;
 import net.dv8tion.jda.api.events.interaction.SlashCommandEvent;
 import net.dv8tion.jda.api.events.message.guild.GuildMessageReceivedEvent;
-import nl.daanh.hiromi.commands.currency.BalanceCommand;
-import nl.daanh.hiromi.commands.currency.BankCommand;
+import nl.daanh.hiromi.commands.fun.BalanceCommand;
+import nl.daanh.hiromi.commands.fun.BankCommand;
 import nl.daanh.hiromi.commands.other.LoadCommand;
 import nl.daanh.hiromi.commands.other.PingCommand;
 import nl.daanh.hiromi.commands.other.SettingsCommand;
